core: reuse the trunks client for the channel usage anchor
The channel usage queue adapter had grown its own copy of the realtime keyspace
knowledge: db index, scan count, a lax trunks:* pattern and a regexp matching
only the brand and company segments. That is the duplication PROVIDER-2483 had
just removed elsewhere, and it would go quietly wrong the next time the key
layout gains a segment.

The occupancy anchor now comes from TrunksClientInterface, so the adapter is
left owning only the event queue and Provider holds no key layout at all.

Also introduces a ChannelUsageEvent value object with fromWire(), replacing the
array shapes that were threaded through the parser, the calculator and the
collector, and moves per-company row building into ChannelUsageRowBuilder.
CollectChannelUsage is down to 108 lines and now only orchestrates: read a
bounded slice, split closed buckets from the open one, delegate, drain.
